interpreter: Split decorators from interpreterbase.py
